requirements.txt
streamlit
tensorflow
Pillow
numpy
gdown
mplfinance
yfinance
pandas